refactor: Rename TripLayout in TripDialogMobile
and split content into TripDialogMobileContent component
